+++ src/react/components/table/XUITable.tsx
@@ -183,6 +183,10 @@ interface BaseProps<RD extends RowData> {
    * have the `onRowClick` click handler applied to it.
    */
   shouldRowClick?: (rowData: RD) => boolean;
+  /**
+   * Sets the table's layout to `fixed`.
+   */
+  useFixedLayout?: boolean;
 }
 
 type Props<RD extends RowData> = BaseProps<RD>;
@@ -245,6 +249,7 @@ class XUITable<RD extends RowData = RowData> extends React.PureComponent<Props<R
       overflowMenuTitle,
       qaHook,
       shouldRowClick,
+      useFixedLayout,
     } = this.props;
 
     const checkedRowIds = Object.keys(checkedIds).filter(key => checkedIds[key]);
@@ -284,7 +289,10 @@ class XUITable<RD extends RowData = RowData> extends React.PureComponent<Props<R
           qaHook={qaHook && `${qaHook}-table`}
           ref={this.tableNode}
           scrollContainerRef={this.wrapperNode}
-          tableClassName={cn(isTruncated && `${tableName}-is-truncated`)}
+          tableClassName={cn(
+            isTruncated && `${tableName}-is-truncated`,
+            useFixedLayout && `${tableName}-fixed-layout`,
+          )}
         >
           {everyColumnHasHeader && (
             <XUITableHead
@@ -394,6 +402,7 @@ class XUITable<RD extends RowData = RowData> extends React.PureComponent<Props<R
     overflowMenuTitle: PropTypes.string,
     qaHook: PropTypes.string,
     shouldRowClick: PropTypes.func,
+    useFixedLayout: PropTypes.bool,
   };
 }